requirements.txt
bcrypt
python-dotenv
flask 
flask_pymongo
Pillow
